Доработал интерфейс, добавил несколько моделей отображения

+++ PizzaDatabase/partner_products.sql
@@ -2,7 +2,7 @@
 (
 	[PartnerProductId] INT NOT NULL PRIMARY KEY IDENTITY,
 	[ProductId] INT NOT NULL,
-	[ProductPartnerPrice] FLOAT NOT NULL, 
+	[ProductPartnerPrice] INT NOT NULL, 
     [ProductCount] INT NOT NULL, 
     FOREIGN KEY ([ProductId]) REFERENCES dbo.products_table([ProductId])
 )
