+++ PizzaInternalSystem/Models/PartnerProductModel.cs
@@ -4,7 +4,7 @@
     {
         public int PartnerProductId { get; set; }
         public int ProductId { get; set; }
-        public float ProductPartnerPrice { get; set; }
+        public int ProductPartnerPrice { get; set; }
         public int ProductCount { get; set; }
     }
 }
